+++ packages/core/src/index.ts
@@ -995,3 +995,4 @@ export * from './constants'
 export { PassiveStrategy, PromiscuousStrategy, SaneDefaults, findPath }
 export type { ChannelsToOpen, ChannelsToClose }
 export type { ProtocolConfig, Network } from './environment'
+export { resolveEnvironment } from './environment'
