+++ packages/cover-traffic/src/index.ts
@@ -1,8 +1,7 @@
 import { PersistedState } from './state'
 import type { PeerData, State } from './state'
 import type { HoprOptions } from '@hoprnet/hopr-core'
-import { resolveEnvironment } from '@hoprnet/hopr-core'
-import Hopr from '@hoprnet/hopr-core'
+import Hopr, { resolveEnvironment } from '@hoprnet/hopr-core'
 import { CoverTrafficStrategy } from './strategy'
 import { ChannelEntry, privKeyToPeerId, PublicKey } from '@hoprnet/hopr-utils'
 import type PeerId from 'peer-id'
